front - Rename DiscussionContentTab to DiscussionTabContent according to component naming
