+++ front/src/app/components/DiscussionTabContent.tsx
